// src/app/core/interfaces/promedioDimensiones.ts
export interface PromedioDimensiones {
    idPromedioDim?: number;
    idTablero?: number;
    lote?: string;
    tipo?: string;
    d1TalentoHumano: number | null;
    d2DireccionamientoEstrategicoPlaneacion: number | null;
    d3GestionResultadosValores: number | null;
    d4EvaluacionResultados: number | null;
    d5InformacionComunicacion: number | null;
    d6GestionConocimiento: number | null;
    d7Control: number | null;
    idi: number | null;
    activo: boolean | null;
}